requirements.txt
openai==1.12.0
python-dotenv==1.0.0
tqdm==4.66.2
elevenlabs==1.54.0
moviepy==1.0.3
openai-whisper==20231117
numpy==1.26.4
torch==2.2.1